修改了MotorDJI库的逻辑
将MotorDJI库的驱动与控制部分解耦了
